Fixed issue with IPv6 addresses found during interop tests
Fixed issue when the URI for an OSCORE context was set to an IPv6 address.
The problem was that the normalizeServerURI method did not normalize IPv6 addresses correctly depending on their structure. This meant getContext and addContext could use different URIs so the context was not found when trying to retrieve it.
I have now made a change to the normalizeServerURI method with some additions for extra normalization of IPv6 addresses.