Changed prints of key information and external AAD serialization
*Added further prints of key information
In RequestEncryptor.java, RequestDecryptor.java, ResponseEncryptor.java and ResponseDecryptor.java
External AAD, ciphertext and plaintext is now printed. Plus full AAD / Encrypt0 structure from EncryptCommon.java

*Changed serialization of external AAD
In OSSerializer.java
It now uses a pre-prepared value with an empty byte array that was in the code since before.
Note that the encoding of the external AAD still has to be fixed to take care of the last parameter properly.

*Added some further comments regarding things TODO