Aligning code with the latest Californium 2.0.x branch
Updates this repository, that has code containing the interop test client
and server, with the latest code from the Californium 2.0.x branch. It is
now aligned with the latest Californium and OSCORE code, and includes the
interop test client and server.

The interop test client and server were not included in the pull requests
done to the official Californium repository. This due to wanting to keep
the commits small and the fact that the code did not really belong in the
official repository. However it is useful to have a repository based on
the latest Californium and OSCORE code that also includes the interop
test client and server.

This code also contains a JUnit-version of most of the interop tests in
OSCoreMessageExchangeTest.java

The code from Californium was taken from the 2.0.x branch at commit:
https://github.com/eclipse/californium/commit/02f7e696a9db5b888be2d1ee940272b60d78b3ad

The interop test client and server tests the interop test specification from Ericsson:
https://ericssonresearch.github.io/OSCOAP/test-spec5.html
https://ericssonresearch.github.io/OSCOAP/